restructured files and added a random command
