+++ main.py
@@ -8,9 +8,8 @@ from discord import app_commands
 from discord.ext import commands
 from dotenv import load_dotenv
 
-# -- IDK it's for the markov babbler commands -- #
+# -- included librarys -- #
 import markov_babbler as Babbler
-import dropbox
 
 load_dotenv()
 
@@ -19,29 +18,9 @@ if BOT_TOKEN is None:
     raise ValueError('Environment variable "BOT_TOKEN" not set.')
 
 # -- Data storage for markov commands -- #
-DROPBOX_KEY = os.getenv('DROPBOX_KEY')
-if DROPBOX_KEY is None:
-    raise ValueError('Environment variable "DROPBOX_KEY" not set.')
-dbx = dropbox.Dropbox(DROPBOX_KEY)
 messages_sent = 0
 need_new_stats = True
-
-def get_cloud_stats():
-    try:
-        metadata, res = dbx.files_download("/lilguys-markov/stats.json")
-        data = res.content
-        
-        stats = json.loads(data.decode('utf-8'))
-        need_new_stats = False
-        return stats
-    except Exception as e:
-        need_new_stats = False
-        return {}
-
-def save_cloud_stats(stats):
-    stats_str = json.dumps(stats)
-    stats_bytes = stats_str.encode('utf-8')
-    dbx.files_upload(stats_bytes, '/lilguys-markov/stats.json', mode=dropbox.files.WriteMode.overwrite)
+stats = {}
 
 intents = discord.Intents.default()
 intents.message_content = True
@@ -56,9 +35,16 @@ async def on_ready():
 @bot.tree.command(name="babble", description="Generate Markov babble")
 @app_commands.describe(sentences="Number of sentences to generate")
 async def babble(interaction: discord.Interaction, sentences: int):
-    stats = get_cloud_stats()
+    global stats
     await interaction.response.send_message(Babbler.babble(stats, sentences))
 
+# -- random command -- #
+@bot.tree.command(name="anagram", description="Figure out if two words are anagrams")
+@app_commands.describe(word1="first word", word2="Second word")
+async def anagram(interaction: discord.Interaction, word1: str, word2: str):
+    angrm = "are" if sorted(word1.lower().strip()) == sorted(word2.lower().strip()) else "are not"
+    await interaction.response.send_message(f"{word1} and {word2} {angrm} anagrams.")
+ 
 @bot.event
 async def on_message(message):
     global messages_sent
@@ -69,10 +55,10 @@ async def on_message(message):
 
     messages_sent += 1
     if need_new_stats:
-        stats = get_cloud_stats()
+        stats = Babbler.get_cloud_stats()
     stats = Babbler.get_stats(message.content, stats)
     if messages_sent >= 15:
-        save_cloud_stats(stats)
+        Babbler.save_cloud_stats(stats)
         need_new_stats = True
 
 bot.run(BOT_TOKEN)
